Adds option to test process death.
Process death enables us to test if the app recovers correctly when the OS kills the app to reclaim resources.
The test consist of sending the app to the background if it is in the foreground, kill the app and then relaunch it. At this point, onCreate is called with savedInstanceState not null.

Adds method to check is app is in the foreground.